+++ extensions/api/control/src/main/java/org/eclipse/dataspaceconnector/api/control/ClientApi.java
@@ -1,0 +1,34 @@
+/*
+ *  Copyright (c) 2022 Bayerische Motoren Werke Aktiengesellschaft (BMW AG)
+ *
+ *  This program and the accompanying materials are made available under the
+ *  terms of the Apache License, Version 2.0 which is available at
+ *  https://www.apache.org/licenses/LICENSE-2.0
+ *
+ *  SPDX-License-Identifier: Apache-2.0
+ *
+ *  Contributors:
+ *       Bayerische Motoren Werke Aktiengesellschaft (BMW AG)
+ *
+ */
+
+package org.eclipse.dataspaceconnector.api.control;
+
+import io.swagger.v3.oas.annotations.OpenAPIDefinition;
+import jakarta.ws.rs.core.Response;
+import org.eclipse.dataspaceconnector.spi.types.domain.contract.negotiation.ContractOfferRequest;
+import org.eclipse.dataspaceconnector.spi.types.domain.transfer.DataRequest;
+
+@OpenAPIDefinition
+public interface ClientApi {
+
+    Response addTransfer(DataRequest dataRequest);
+
+    Response initiateNegotiation(ContractOfferRequest contractOffer);
+
+    Response getNegotiationById(String id);
+
+    Response getNegotiationStateById(String id);
+
+    Response getAgreementById(String id);
+}