+++ extensions/api/data-management/transferprocess/src/main/java/org/eclipse/dataspaceconnector/api/datamanagement/contractdefinition/TransferProcessApiController.java
@@ -9,6 +9,7 @@
  *
  *  Contributors:
  *       Microsoft Corporation - initial API and implementation
+ *       Bayerische Motoren Werke Aktiengesellschaft (BMW AG)
  *
  */
 
@@ -35,7 +36,7 @@ import static java.lang.String.format;
 @Consumes({ MediaType.APPLICATION_JSON })
 @Produces({ MediaType.APPLICATION_JSON })
 @Path("/transferprocess")
-public class TransferProcessApiController {
+public class TransferProcessApiController implements TransferProcessApi {
     private final Monitor monitor;
 
     public TransferProcessApiController(Monitor monitor) {
@@ -43,11 +44,12 @@ public class TransferProcessApiController {
     }
 
     @GET
-    public List<TransferProcessDto> getAllContractDefinitions(@QueryParam("offset") Integer offset,
-                                                              @QueryParam("limit") Integer limit,
-                                                              @QueryParam("filter") String filterExpression,
-                                                              @QueryParam("sort") SortOrder sortOrder,
-                                                              @QueryParam("sortField") String sortField) {
+    @Override
+    public List<TransferProcessDto> getAllTransferProcesses(@QueryParam("offset") Integer offset,
+                                                            @QueryParam("limit") Integer limit,
+                                                            @QueryParam("filter") String filterExpression,
+                                                            @QueryParam("sort") SortOrder sortOrder,
+                                                            @QueryParam("sortField") String sortField) {
         var spec = QuerySpec.Builder.newInstance()
                 .offset(offset)
                 .limit(limit)
@@ -62,6 +64,7 @@ public class TransferProcessApiController {
 
     @GET
     @Path("{id}")
+    @Override
     public TransferProcessDto getTransferProcess(@PathParam("id") String id) {
         monitor.debug(format("get TransferProcess with ID %s", id));
 
@@ -70,6 +73,7 @@ public class TransferProcessApiController {
 
     @GET
     @Path("{id}/state")
+    @Override
     public String getTransferProcessState(@PathParam("id") String id) {
         monitor.debug(format("get TransferProcess State with ID %s", id));
 
@@ -78,12 +82,14 @@ public class TransferProcessApiController {
 
     @POST
     @Path("{id}/cancel")
+    @Override
     public void cancelTransferProcess(@PathParam("id") String id) {
         monitor.debug("Cancelling TransferProcess with ID " + id);
     }
 
     @POST
     @Path("{id}/deprovision")
+    @Override
     public void deprovisionTransferProcess(@PathParam("id") String id) {
         monitor.debug(format("Attempting to deprovision TransferProcess with id %s", id));
     }