+++ extensions/api/observability/src/main/java/org/eclipse/dataspaceconnector/api/observability/ObservabilityApiController.java
@@ -1,5 +1,5 @@
 /*
- *  Copyright (c) 2021 Daimler TSS GmbH
+ *  Copyright (c) 2021 - 2022 Daimler TSS GmbH
  *
  *  This program and the accompanying materials are made available under the
  *  terms of the Apache License, Version 2.0 which is available at
@@ -10,6 +10,7 @@
  *  Contributors:
  *       Daimler TSS GmbH - Initial API and Implementation
  *       Fraunhofer Institute for Software and Systems Engineering - add negotiation endpoint
+ *       Bayerische Motoren Werke Aktiengesellschaft (BMW AG)
  *
  */
 
@@ -29,8 +30,7 @@ import org.eclipse.dataspaceconnector.spi.system.health.HealthStatus;
 @Consumes({ MediaType.APPLICATION_JSON })
 @Produces({ MediaType.APPLICATION_JSON })
 @Path("/check")
-public class ObservabilityApiController {
-
+public class ObservabilityApiController implements ObservabilityApi {
 
     private final HealthCheckService healthCheckService;
 
@@ -40,6 +40,7 @@ public class ObservabilityApiController {
 
     @GET
     @Path("health")
+    @Override
     public Response checkHealth() {
         var status = healthCheckService.getStartupStatus();
         return createResponse(status);
@@ -47,6 +48,7 @@ public class ObservabilityApiController {
 
     @GET
     @Path("liveness")
+    @Override
     public Response getLiveness() {
         var status = healthCheckService.isLive();
         return createResponse(status);
@@ -55,6 +57,7 @@ public class ObservabilityApiController {
 
     @GET
     @Path("readiness")
+    @Override
     public Response getReadiness() {
         var status = healthCheckService.isReady();
         return createResponse(status);
@@ -62,6 +65,7 @@ public class ObservabilityApiController {
 
     @GET
     @Path("startup")
+    @Override
     public Response getStartup() {
         var status = healthCheckService.getStartupStatus();
         return createResponse(status);
@@ -71,7 +75,6 @@ public class ObservabilityApiController {
         return status.isHealthy() ?
                 Response.ok().entity(status).build() :
                 Response.status(503).entity(status).build();
-
     }
 
 }